+++ org.eclim.core/java/org/eclim/plugin/core/command/project/ProjectNatureAddCommand.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2005 - 2011  Eric Van Dewoestine
+ * Copyright (C) 2005 - 2012  Eric Van Dewoestine
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -51,7 +51,8 @@ import org.eclipse.core.runtime.NullProgressMonitor;
   name = "project_nature_add",
   options =
     "REQUIRED p project ARG," +
-    "REQUIRED n nature ARG"
+    "REQUIRED n nature ARG," +
+    "OPTIONAL a args ANY"
 )
 public class ProjectNatureAddCommand
   extends AbstractCommand