+++ doc/content/changes.rst
@@ -16,6 +16,24 @@
 History of Changes
 ==================
 
+.. _2.2.1:
+
+2.2.1 (Aug. XX, 2012)
+----------------------
+
+Android:
+  - Eclim now has support for :ref:`creating android projects
+    <gettingstarted-android>`.
+
+.. _2.2.0:
+
+2.2.0 (Aug. 07, 2012)
+----------------------
+
+Eclipse:
+  - Eclim now requires Java 6 or later.
+  - Eclim now requires the latest version of eclipse (Juno, 4.2).
+
 .. _1.7.7:
 
 1.7.7 (Aug. 07, 2012)